fix mypy issue and use same code everywhere
