+++ antarest/study/storage/utils.py
@@ -17,7 +17,6 @@ import os
 import re
 import shutil
 import tempfile
-import time
 from datetime import datetime, timedelta
 from pathlib import Path
 from typing import List, Optional, Sequence, cast
@@ -59,7 +58,7 @@ from antarest.study.model import (
     StudyMetadataDTO,
 )
 from antarest.study.storage.rawstudy.model.filesystem.config.model import Mode
-from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy, StudyFactory
+from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.rawstudy.model.filesystem.root.filestudytree import FileStudyTree
 from antarest.study.storage.rawstudy.model.helpers import FileStudyHelpers
 
@@ -358,57 +357,6 @@ def get_start_date(
         first_week_size=first_week_size,
         level=level,
     )
-
-
-def export_study_flat(
-    study_dir: Path,
-    dest: Path,
-    study_factory: StudyFactory,
-    outputs: bool = True,
-    output_list_filter: Optional[List[str]] = None,
-    denormalize: bool = True,
-    output_src_path: Optional[Path] = None,
-    is_study_managed: bool = True,
-) -> None:
-    start_time = time.time()
-
-    output_src_path = output_src_path or study_dir / "output"
-    output_dest_path = dest / "output"
-
-    def ignore_outputs(directory: str, _: Sequence[str]) -> Sequence[str]:
-        return ["output"] if str(directory) == str(study_dir) else []
-
-    shutil.copytree(src=study_dir, dst=dest, ignore=ignore_outputs)
-
-    if outputs and output_src_path.exists():
-        if output_list_filter is None:
-            # Retrieve all directories or ZIP files without duplicates
-            output_list_filter = list(
-                {f.with_suffix("").name for f in output_src_path.iterdir() if f.is_dir() or f.suffix == ".zip"}
-            )
-        # Copy each folder or uncompress each ZIP file to the destination dir.
-        shutil.rmtree(output_dest_path, ignore_errors=True)
-        output_dest_path.mkdir()
-        for output in output_list_filter:
-            zip_path = output_src_path / f"{output}.zip"
-            if zip_path.exists():
-                with ZipFile(zip_path) as zf:
-                    zf.extractall(output_dest_path / output)
-            else:
-                shutil.copytree(
-                    src=output_src_path / output,
-                    dst=output_dest_path / output,
-                )
-
-    stop_time = time.time()
-    duration = "{:.3f}".format(stop_time - start_time)
-    with_outputs = "with outputs" if outputs else "without outputs"
-    logger.info(f"Study '{study_dir}' exported ({with_outputs}, flat mode) in {duration}s")
-    study = study_factory.create_from_fs(dest, is_study_managed, "", use_cache=False)
-    if denormalize:
-        study.tree.denormalize()
-        duration = "{:.3f}".format(time.time() - stop_time)
-        logger.info(f"Study '{study_dir}' denormalized in {duration}s")
 
 
 def is_folder_safe(workspace: WorkspaceConfig, folder: str) -> bool:
